Updating the API functions and updating some helpers

+++ Helpers/eventHelper.js
@@ -15,7 +15,7 @@ class EventHelper {
             return updateEvent(event);
         }
 
-        else {
+        else if (method === "DELETE") {
             return deleteEvent(event);
         }
     }
